moved AnimateAddToBoard and AnimateClearLines to Animation folder. created empty Animation file + tests
