+++ frontend/static/js/Command/Animation/AnimateAddToBoard.js
