+++ frontend/static/js/Command/Animation/AnimateClearLines.js
